// src/Umbraco.Cms.Search.Core.Client/Client/src/settings/api/types.gen.ts
export type DirectionModel = 'Ascending' | 'Descending';

export type DocumentModel = {
    id: string;
    objectType: UmbracoObjectTypesModel;
    name?: null | string;
    icon?: null | string;
};

export type FacetModel = {
    fieldName: string;
};

export type FacetResultModel = {
    fieldName: string;
    values: Array<FacetValueModel>;
};

export type FacetValueModel = {
    count: number;
};

export type FilterModel = {
    fieldName: string;
    negate: boolean;
};

export type HealthStatusModel = 'Healthy' | 'Rebuilding' | 'Corrupted' | 'Empty' | 'Unknown';

export type IndexModel = {
    indexAlias: string;
    providerName: string;
    documentCount: number;
    healthStatus: HealthStatusModel;
};

export type PagedIndexModel = {
    total: number;
    items: Array<IndexModel>;
};

export type SearchRequestModel = {
    indexAlias: string;
    query?: null | string;
    filters?: null | Array<FilterModel>;
    facets?: null | Array<FacetModel>;
    sorters?: null | Array<SorterModel>;
    culture?: null | string;
    segment?: null | string;
};

export type SearchResultModel = {
    total: number;
    documents: Array<DocumentModel>;
    facets: Array<FacetResultModel>;
};

export type SorterModel = {
    fieldName: string;
    direction: DirectionModel;
};